优化计算graph点的方式，转换为double

+++ lib/entity/draw_graph_entity.dart
@@ -6,7 +6,7 @@ enum DrawGraphType {
 }
 
 class DrawGraphValue {
-  int index;
+  double index;
   double price;
 
   DrawGraphValue(
